Add constructs for safe handling of PII. (#28)
Add constructs for safe handling of PII.

This change centralizes PII handling with a series of wrapper classes.
These classes selectively store or redact data based upon the
`--collect_pii` flag. This lets the actual scraping code just grab
everything, throw it in the PII wrappers and trust that they'll Do The
Right Thing, without materially changing the usage of the PII wrapped
data. This is accomplished by subclassing the data we wish to
encapsulate. A `Pii.String` is a subclass of `str`, for example.

PII handling subclasses are cool and all, but they aren't really useful
if they don't get consistently used. To that end, this change alters our
`Record` dataclass to include complete type hints for all fields and
ensures the field values match the type at runtime. Attempting to shove
a normal string into a `Pii.String`-typed field will result in a
`TypeError`, for example.

While we're here, define common `Record` and `Charge` definitions
to call out optional fields and expose builder objects for them. Given
the number of fields these have, this is less error prone than setting a
bunch of individual variables and plugging them into our dataclass
constructors at the end. These common dataclasses encompass the
minimal fields we wish to collect and may be subclassed to add
additional fields if that's desired. Since the FL/Bay/Court scraper has
been populating additional fields for a while, have it use a subclassed
BenchmarkRecord (with accompanying builder) to collect its
scrapings.