Removal of Scraper Code and Notes Added to README.MD
Per the Leadership Meeting discussion on 01/06/2020, I have removed the scraper code that allows for the bypass of CAPTCHAS to further adhere to our own core values. This could be used to violate a website's or organization's Terms of Service which we do not condone or endorse when collecting data for the Police Data Accessibility Project.