Update my fork (#27)
* UUID Fix w/ no dashes

Changed generated UUIDs to remove dashes, in-line with the dolt database schema

* flesh out schema.json and added guide

* Updated schema with dict

Forward-thinking, agencies could have multiple different sets of data. (incident reports, booking reports, traffic stops). We would want each type of data in a separate sub-directory and then in the schema file tell the ETL script what it is.

* increased indent for `data`

* address agency - dataset issue

We will need to branch agency and datasets off of a one-to-one model because a single agency can have multiple different datasets.

* Adds Brea

* Upgraded ETL scripts

New etl scripts interface with both the `datasets` repo and the `data-intake` repo.

Currently, it will search for an agency & dataset, and add a new dataset if missing.

However, there are no checks on the `data-intake`, for now it just simply inserts all good data.

Co-authored-by: Eric Turner <me@ericturner.it>
Co-authored-by: Eric Turner <turner.eric7@gmail.com>